requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.0

# OpenAI Client (for ModelScope API)
openai==1.10.0

# Vector Database
chromadb==0.4.22
numpy<2.0.0

# Data Validation
pydantic==2.5.3

# CORS Support
python-jose[cryptography]==3.3.0
